refactor: 팔로워버튼과 메뉴 컴포넌트 파일로 분리
